Adding current_tx

+++ library/include/tnpheap.h
@@ -3,7 +3,7 @@ extern "C" {
 #endif
 #include <unistd.h>
 #include <signal.h>
-int current_tx;
+__u64 current_tx;
 __u64 tnpheap_start_tx(int npheap_dev, int tnpheap_dev);
 int tnpheap_commit(int npheap_dev, int tnpheap_dev);
 void *tnpheap_alloc(int npheap_dev, int tnpheap_dev, __u64 offset, __u64 size);
